add polars LazyFrame generic type, element-wise checks, add docs (#1521)
* add polars LazyFrame generic type, add docs


* readme, makefile fixes


* add support for pl.DataFrame


* fix failure case collection logic


* update docs


* add custom check docs


* add polars elementwise check support


* update docs


* add NotImplementedError for polars strategies


* update docs


---------
