+++ docs/source/index.rst
@@ -78,7 +78,7 @@ This is useful in production-critical data pipelines or reproducible research
 settings. With ``pandera``, you can:
 
 #. Define a schema once and use it to validate :ref:`different dataframe types <supported-dataframe-libraries>`
-   including `pandas <http://pandas.pydata.org>`_, `dask <https://dask.org/>`_,
+   including `pandas <http://pandas.pydata.org>`_, `polars <https://docs.pola.rs/>`, `dask <https://dask.org/>`_,
    `modin <https://modin.readthedocs.io/>`_, and
    `pyspark.pandas <https://spark.apache.org/docs/3.2.0/api/python/user_guide/pandas_on_spark/index.html>`_.
 #. :ref:`Check<checks>` the types and properties of columns in a
@@ -137,6 +137,7 @@ Installing additional functionality:
       pip install pandera[modin-ray]   # validate modin dataframes with ray
       pip install pandera[modin-dask]  # validate modin dataframes with dask
       pip install pandera[geopandas]   # validate geopandas geodataframes
+      pip install pandera[polars]      # validate polars dataframes
 
 .. tabbed:: conda
 
@@ -153,6 +154,7 @@ Installing additional functionality:
       conda install -c conda-forge pandera-modin-ray   # validate modin dataframes with ray
       conda install -c conda-forge pandera-modin-dask  # validate modin dataframes with dask
       conda install -c conda-forge pandera-geopandas   # validate geopandas geodataframes
+      conda install -c conda-forge pandera-polars      # validate polars dataframes
 
 Quick Start
 -----------
