--- docs/source/koalas.rst
@@ -1,11 +1,0 @@
-.. currentmodule:: pandera
-
-.. _scaling_koalas:
-
-Data Validation with Koalas
-===========================
-
-.. note::
-
-    Koalas has been deprecated since version *0.10.0*. Please refer to the
-    :ref:`pyspark page <scaling_pyspark>` for validating pyspark dataframes.
