+++ pandera/api/polars/container.py
@@ -6,6 +6,7 @@ from typing import Optional
 import polars as pl
 
 from pandera.api.pandas.container import DataFrameSchema as _DataFrameSchema
+from pandera.api.polars.types import PolarsCheckObjects
 from pandera.dtypes import DataType
 from pandera.engines import polars_engine
 
@@ -30,17 +31,21 @@ class DataFrameSchema(_DataFrameSchema):
 
     def validate(
         self,
-        check_obj: pl.LazyFrame,
+        check_obj: PolarsCheckObjects,
         head: Optional[int] = None,
         tail: Optional[int] = None,
         sample: Optional[int] = None,
         random_state: Optional[int] = None,
         lazy: bool = False,
         inplace: bool = False,
-    ) -> pl.LazyFrame:
+    ) -> PolarsCheckObjects:
         """Validate a polars DataFrame against the schema."""
+        is_dataframe = isinstance(check_obj, pl.DataFrame)
 
-        return self.get_backend(check_obj).validate(
+        if is_dataframe:
+            check_obj = check_obj.lazy()
+
+        output = self.get_backend(check_obj).validate(
             check_obj=check_obj,
             schema=self,
             head=head,
@@ -50,6 +55,11 @@ class DataFrameSchema(_DataFrameSchema):
             lazy=lazy,
             inplace=inplace,
         )
+
+        if is_dataframe:
+            output = output.collect()
+
+        return output
 
     @property
     def dtype(
@@ -62,3 +72,34 @@ class DataFrameSchema(_DataFrameSchema):
     def dtype(self, value) -> None:
         """Set the pandas dtype property."""
         self._dtype = polars_engine.Engine.dtype(value) if value else None
+
+    def strategy(
+        self, *, size: Optional[int] = None, n_regex_columns: int = 1
+    ):
+        """Create a ``hypothesis`` strategy for generating a DataFrame.
+
+        :param size: number of elements to generate
+        :param n_regex_columns: number of regex columns to generate.
+        :returns: a strategy that generates pandas DataFrame objects.
+
+        .. warning::
+
+           This method is not implemented in the polars backend.
+        """
+        raise NotImplementedError(
+            "Data synthesis is not supported in with polars schemas."
+        )
+
+    def example(self, size: Optional[int] = None, n_regex_columns: int = 1):
+        """Generate an example of a particular size.
+
+        :param size: number of elements in the generated DataFrame.
+        :returns: pandas DataFrame object.
+
+        .. warning::
+
+           This method is not implemented in polars backend.
+        """
+        raise NotImplementedError(
+            "Data synthesis is not supported in with polars schemas."
+        )
