+++ pandera/backends/polars/builtin_checks.py
@@ -24,7 +24,7 @@ def equal_to(data: PolarsData, value: Any) -> pl.LazyFrame:
     :param value: values in this polars data structure must be
         equal to this value.
     """
-    return data.dataframe.select(pl.col(data.key).eq(value))
+    return data.lazyframe.select(pl.col(data.key).eq(value))
 
 
 @register_builtin_check(
@@ -38,7 +38,7 @@ def not_equal_to(data: PolarsData, value: Any) -> pl.LazyFrame:
                 to access the dataframe is "dataframe" and column name using "key".
     :param value: This value must not occur in the checked
     """
-    return data.dataframe.select(pl.col(data.key).ne(value))
+    return data.lazyframe.select(pl.col(data.key).ne(value))
 
 
 @register_builtin_check(
@@ -55,7 +55,7 @@ def greater_than(data: PolarsData, min_value: Any) -> pl.LazyFrame:
     :param min_value: Lower bound to be exceeded. Must be
             a type comparable to the dtype of the series datatype of Polars
     """
-    return data.dataframe.select(pl.col(data.key).gt(min_value))
+    return data.lazyframe.select(pl.col(data.key).gt(min_value))
 
 
 @register_builtin_check(
@@ -70,7 +70,7 @@ def greater_than_or_equal_to(data: PolarsData, min_value: Any) -> pl.LazyFrame:
     :param min_value: Allowed minimum value for values of a series. Must be
             a type comparable to the dtype of the series datatype of Polars
     """
-    return data.dataframe.select(pl.col(data.key).ge(min_value))
+    return data.lazyframe.select(pl.col(data.key).ge(min_value))
 
 
 @register_builtin_check(
@@ -85,7 +85,7 @@ def less_than(data: PolarsData, max_value: Any) -> pl.LazyFrame:
     :param max_value: All elements of a series must be strictly smaller
         than this. Must be a type comparable to the dtype of the series datatype of Polars
     """
-    return data.dataframe.select(pl.col(data.key).lt(max_value))
+    return data.lazyframe.select(pl.col(data.key).lt(max_value))
 
 
 @register_builtin_check(
@@ -100,7 +100,7 @@ def less_than_or_equal_to(data: PolarsData, max_value: Any) -> pl.LazyFrame:
     :param max_value: Upper bound not to be exceeded. Must be a type comparable to the dtype of the
     series datatype of Polars
     """
-    return data.dataframe.select(pl.col(data.key).le(max_value))
+    return data.lazyframe.select(pl.col(data.key).le(max_value))
 
 
 @register_builtin_check(
@@ -135,7 +135,7 @@ def in_range(
     is_in_min = col.ge(min_value) if include_min else col.gt(min_value)
     is_in_max = col.le(max_value) if include_max else col.lt(max_value)
 
-    return data.dataframe.select(is_in_min.and_(is_in_max))
+    return data.lazyframe.select(is_in_min.and_(is_in_max))
 
 
 @register_builtin_check(
@@ -155,7 +155,7 @@ def isin(data: PolarsData, allowed_values: Iterable) -> pl.LazyFrame:
                 to access the dataframe is "dataframe" and column name using "key".
     :param allowed_values: The set of allowed values. May be any iterable.
     """
-    return data.dataframe.select(pl.col(data.key).is_in(allowed_values))
+    return data.lazyframe.select(pl.col(data.key).is_in(allowed_values))
 
 
 @register_builtin_check(
@@ -174,7 +174,7 @@ def notin(data: PolarsData, forbidden_values: Iterable) -> pl.LazyFrame:
     :param forbidden_values: The set of values which should not occur. May
         be any iterable.
     """
-    return data.dataframe.select(
+    return data.lazyframe.select(
         pl.col(data.key).is_in(forbidden_values).not_()
     )
 
@@ -195,7 +195,7 @@ def str_matches(
     pattern = pattern.pattern if isinstance(pattern, re.Pattern) else pattern
     if not pattern.startswith("^"):
         pattern = f"^{pattern}"
-    return data.dataframe.select(
+    return data.lazyframe.select(
         pl.col(data.key).str.contains(pattern=pattern)
     )
 
@@ -215,7 +215,7 @@ def str_contains(
     """
 
     pattern = pattern.pattern if isinstance(pattern, re.Pattern) else pattern
-    return data.dataframe.select(
+    return data.lazyframe.select(
         pl.col(data.key).str.contains(pattern=pattern, literal=False)
     )
 
@@ -231,7 +231,7 @@ def str_startswith(data: PolarsData, string: str) -> pl.LazyFrame:
     :param string: String all values should start with
     """
 
-    return data.dataframe.select(pl.col(data.key).str.starts_with(string))
+    return data.lazyframe.select(pl.col(data.key).str.starts_with(string))
 
 
 @register_builtin_check(error="str_endswith('{string}')")
@@ -242,7 +242,7 @@ def str_endswith(data: PolarsData, string: str) -> pl.LazyFrame:
                 to access the dataframe is "dataframe" and column name using "key".
     :param string: String all values should end with
     """
-    return data.dataframe.select(pl.col(data.key).str.ends_with(string))
+    return data.lazyframe.select(pl.col(data.key).str.ends_with(string))
 
 
 @register_builtin_check(
@@ -273,7 +273,7 @@ def str_length(
     else:
         expr = n_chars.is_between(min_value, max_value)
 
-    return data.dataframe.select(expr)
+    return data.lazyframe.select(expr)
 
 
 @register_builtin_check(
@@ -292,5 +292,5 @@ def unique_values_eq(data: PolarsData, values: Iterable) -> bool:
     """
 
     return (
-        set(data.dataframe.collect().get_column(data.key).unique()) == values
+        set(data.lazyframe.collect().get_column(data.key).unique()) == values
     )
