+++ pandera/backends/polars/checks.py
@@ -10,10 +10,7 @@ from pandera.api.base.checks import CheckResult
 from pandera.api.checks import Check
 from pandera.api.polars.types import PolarsData
 from pandera.backends.base import BaseCheckBackend
-from pandera.backends.polars.constants import (
-    CHECK_OUTPUT_KEY,
-    FAILURE_CASE_KEY,
-)
+from pandera.backends.polars.constants import CHECK_OUTPUT_KEY
 
 
 class PolarsCheckBackend(BaseCheckBackend):
@@ -46,19 +43,24 @@ class PolarsCheckBackend(BaseCheckBackend):
 
     def apply(self, check_obj: PolarsData):
         """Apply the check function to a check object."""
-        out = self.check_fn(check_obj)
+        if self.check.element_wise:
+            out = check_obj.lazyframe.with_columns(
+                pl.col(check_obj.key or "*").map_elements(self.check_fn)
+            )
+        else:
+            out = self.check_fn(check_obj)
 
         if isinstance(out, bool):
             return out
 
         if len(out.columns) > 1:
             # for checks that return a boolean dataframe, reduce to a single
-            # boolean column
+            # boolean column.
             out = out.select(
                 pl.fold(
                     acc=pl.lit(True),
                     function=lambda acc, x: acc & x,
-                    exprs=pl.col(pl.Boolean),
+                    exprs=pl.col("*"),
                 ).alias(CHECK_OUTPUT_KEY)
             )
         else:
@@ -81,14 +83,12 @@ class PolarsCheckBackend(BaseCheckBackend):
     ) -> CheckResult:
         """Postprocesses the result of applying the check function."""
         passed = check_output.select([pl.col(CHECK_OUTPUT_KEY).all()])
-        failure_cases = check_obj.dataframe.with_context(check_output).filter(
+        failure_cases = check_obj.lazyframe.with_context(check_output).filter(
             pl.col(CHECK_OUTPUT_KEY).not_()
         )
 
-        if len(failure_cases.columns) == 1 and check_obj.key is not None:
-            failure_cases = failure_cases.rename(
-                {check_obj.key: FAILURE_CASE_KEY}
-            ).select(FAILURE_CASE_KEY)
+        if check_obj.key is not None:
+            failure_cases = failure_cases.select(check_obj.key)
 
         return CheckResult(
             check_output=check_output,
