+++ pandera/polars.py
@@ -1,9 +1,16 @@
 """A flexible and expressive polars validation library for Python."""
 # pylint: disable=unused-import
+from pandera import errors
 from pandera.api.checks import Check
-from pandera.api.dataframe.model_components import Field
+from pandera.api.dataframe.model_components import (
+    Field,
+    check,
+    dataframe_check,
+)
 from pandera.api.polars.components import Column
 from pandera.api.polars.container import DataFrameSchema
 from pandera.api.polars.model import DataFrameModel
+from pandera.api.polars.types import PolarsData
+from pandera.decorators import check_input, check_io, check_output, check_types
 
 import pandera.backends.polars
