+++ tests/polars/test_polars_container.py
@@ -14,13 +14,19 @@ from pandera.polars import Column, DataFrameSchema
 
 @pytest.fixture
 def ldf_basic():
-    """Basic polars lazy dataframe fixture."""
+    """Basic polars LazyFrame fixture."""
     return pl.DataFrame(
         {
             "string_col": ["0", "1", "2"],
             "int_col": [0, 1, 2],
         }
     ).lazy()
+
+
+@pytest.fixture
+def df_basic(ldf_basic):
+    """Basic polars DataFrame fixture."""
+    return ldf_basic.collect()
 
 
 @pytest.fixture
@@ -82,18 +88,20 @@ def ldf_schema_with_regex_option():
     )
 
 
-def test_basic_polars_lazy_dataframe(ldf_basic, ldf_schema_basic):
+def test_basic_polars_lazyframe(ldf_basic, ldf_schema_basic):
     """Test basic polars lazy dataframe."""
     query = ldf_basic.pipe(ldf_schema_basic.validate)
-    df = query.collect()
+    validated_df = query.collect()
     assert isinstance(query, pl.LazyFrame)
-    assert isinstance(df, pl.DataFrame)
+    assert isinstance(validated_df, pl.DataFrame)
+
+    df = ldf_basic.collect()
+    validated_df = df.pipe(ldf_schema_basic.validate)
+    assert isinstance(validated_df, pl.DataFrame)
 
 
 @pytest.mark.parametrize("lazy", [False, True])
-def test_basic_polars_lazy_dataframe_dtype_error(
-    lazy, ldf_basic, ldf_schema_basic
-):
+def test_basic_polars_lazyframe_dtype_error(lazy, ldf_basic, ldf_schema_basic):
     """Test basic polars lazy dataframe."""
     query = ldf_basic.with_columns(pl.col("int_col").cast(pl.Int32))
 
@@ -103,7 +111,7 @@ def test_basic_polars_lazy_dataframe_dtype_error(
         query.pipe(ldf_schema_basic.validate, lazy=lazy)
 
 
-def test_basic_polars_lazy_dataframe_check_error(
+def test_basic_polars_lazyframe_check_error(
     ldf_basic,
     ldf_schema_with_check,
 ):
@@ -226,13 +234,13 @@ def test_column_values_are_unique(ldf_basic, ldf_schema_basic):
 
 def test_dataframe_level_checks():
     def custom_check(data: PolarsData):
-        return data.dataframe.select(pl.col("*").eq(0))
+        return data.lazyframe.select(pl.col("*").eq(0))
 
     schema = DataFrameSchema(
         columns={"a": Column(pl.Int64), "b": Column(pl.Int64)},
         checks=[
             pa.Check(custom_check),
-            pa.Check(lambda d: d.dataframe.select(pl.col("*").eq(0))),
+            pa.Check(lambda d: d.lazyframe.select(pl.col("*").eq(0))),
         ],
     )
     ldf = pl.DataFrame({"a": [0, 0, 1, 1], "b": [0, 1, 0, 1]}).lazy()
@@ -391,3 +399,27 @@ def test_report_duplicates(arg):
         )
     ):
         DataFrameSchema(report_duplicates=arg)
+
+
+def test_lazy_validation_errors():
+
+    schema = DataFrameSchema(
+        {
+            "a": Column(int),
+            "b": Column(str, C.isin([*"abc"])),
+            "c": Column(float, [C.ge(0.0), C.le(1.0)]),
+        }
+    )
+
+    invalid_lf = pl.LazyFrame(
+        {
+            "a": pl.Series(["1", "2", "3"], dtype=pl.Utf8),  # 1 dtype error
+            "b": ["d", "e", "f"],  # 3 value errors
+            "c": [0.0, 1.1, -0.1],  # 2 value errors
+        }
+    )
+
+    try:
+        schema.validate(invalid_lf, lazy=True)
+    except pa.errors.SchemaErrors as exc:
+        assert exc.failure_cases.shape[0] == 6
